+++ com.omnicon.prm.service/domain/ResourceRQ.go
@@ -10,11 +10,11 @@ type CreateResourceRQ struct {
 }
 
 type DeleteResourceRQ struct {
-	ID int64
+	ID int
 }
 
 type GetResourcesRQ struct {
-	ID            int64
+	ID            int
 	Name          string
 	LastName      string
 	Email         string
@@ -24,9 +24,9 @@ type GetResourcesRQ struct {
 }
 
 type SetResourceToProjectRQ struct {
-	ID         int64
-	ProjectId  int64
-	ResourceId int64
+	ID         int
+	ProjectId  int
+	ResourceId int
 	StartDate  string
 	EndDate    string
 	Lead       bool
@@ -35,9 +35,9 @@ type SetResourceToProjectRQ struct {
 }
 
 type GetResourcesToProjectsRQ struct {
-	ID           int64
-	ProjectId    int64
-	ResourceId   int64
+	ID           int
+	ProjectId    int
+	ResourceId   int
 	ProjectName  string
 	ResourceName string
 	StartDate    string
@@ -47,13 +47,13 @@ type GetResourcesToProjectsRQ struct {
 }
 
 type DeleteResourceToProjectRQ struct {
-	ProjectId  int64
-	ResourceId int64
-	ID         int64
+	ProjectId  int
+	ResourceId int
+	ID         int
 }
 
 type UpdateResourceRQ struct {
-	ID            int64
+	ID            int
 	Name          string
 	LastName      string
 	Email         string
@@ -63,5 +63,5 @@ type UpdateResourceRQ struct {
 }
 
 type GetSkillByResourceRQ struct {
-	ID int64
+	ID int
 }
